Internal #861: Refactor Window Aggregator
Move the window Aggregator base class to separate files.
